+++ thrust/iterator/iterator_categories.h
@@ -32,6 +32,8 @@
 #pragma once
 
 #include <thrust/detail/config.h>
+#include <thrust/iterator/detail/iterator_category_with_system_and_traversal.h>
+#include <thrust/iterator/detail/iterator_traversal_tags.h>
 
 // #include this for stl's iterator tags
 #include <iterator>
@@ -58,7 +60,13 @@ namespace thrust
  *  input_host_iterator_tag, output_host_iterator_tag, forward_host_iterator_tag,
  *  bidirectional_host_iterator_tag, random_access_host_iterator_tag
  */
-struct input_device_iterator_tag {};
+struct input_device_iterator_tag
+  : thrust::detail::iterator_category_with_system_and_traversal<
+      std::input_iterator_tag,
+      thrust::device_system_tag,
+      thrust::single_pass_traversal_tag
+    >
+{};
 
 /*! \p output_device_iterator_tag is an empty class: it has no member functions,
  *  member variables, or nested types. It is used solely as a "tag": a
@@ -71,7 +79,13 @@ struct input_device_iterator_tag {};
  *  input_host_iterator_tag, output_host_iterator_tag, forward_host_iterator_tag,
  *  bidirectional_host_iterator_tag, random_access_host_iterator_tag
  */
-struct output_device_iterator_tag {};
+struct output_device_iterator_tag
+  : thrust::detail::iterator_category_with_system_and_traversal<
+      std::output_iterator_tag,
+      thrust::device_system_tag,
+      thrust::single_pass_traversal_tag
+    >
+{};
 
 /*! \p forward_device_iterator_tag is an empty class: it has no member functions,
  *  member variables, or nested types. It is used solely as a "tag": a
@@ -84,7 +98,13 @@ struct output_device_iterator_tag {};
  *  input_host_iterator_tag, output_host_iterator_tag, forward_host_iterator_tag,
  *  bidirectional_host_iterator_tag, random_access_host_iterator_tag
  */
-struct forward_device_iterator_tag : public input_device_iterator_tag {};
+struct forward_device_iterator_tag
+  : thrust::detail::iterator_category_with_system_and_traversal<
+      std::forward_iterator_tag,
+      thrust::device_system_tag,
+      thrust::forward_traversal_tag
+    >
+{};
 
 /*! \p bidirectional_device_iterator_tag is an empty class: it has no member
  *  functions, member variables, or nested types. It is used solely as a "tag": a
@@ -97,7 +117,13 @@ struct forward_device_iterator_tag : public input_device_iterator_tag {};
  *  input_host_iterator_tag, output_host_iterator_tag, forward_host_iterator_tag,
  *  bidirectional_host_iterator_tag, random_access_host_iterator_tag
  */
-struct bidirectional_device_iterator_tag : public forward_device_iterator_tag {};
+struct bidirectional_device_iterator_tag
+  : thrust::detail::iterator_category_with_system_and_traversal<
+      std::bidirectional_iterator_tag,
+      thrust::device_system_tag,
+      thrust::bidirectional_traversal_tag
+    >
+{};
 
 /*! \p random_access_device_iterator_tag is an empty class: it has no member
  *  functions, member variables, or nested types. It is used solely as a "tag": a
@@ -110,7 +136,13 @@ struct bidirectional_device_iterator_tag : public forward_device_iterator_tag {}
  *  input_host_iterator_tag, output_host_iterator_tag, forward_host_iterator_tag,
  *  bidirectional_host_iterator_tag, random_access_host_iterator_tag
  */
-struct random_access_device_iterator_tag : public bidirectional_device_iterator_tag {};
+struct random_access_device_iterator_tag
+  : thrust::detail::iterator_category_with_system_and_traversal<
+      std::random_access_iterator_tag,
+      thrust::device_system_tag,
+      thrust::random_access_traversal_tag
+    >
+{};
 
 /*! \p input_host_iterator_tag is an empty class: it has no member
  *  functions, member variables, or nested types. It is used solely as a "tag": a
